OSGi bundle changes.  Move classes to avoid package namespace overlap with Spray bundles.
