+++ project/plugins.sbt
@@ -1,1 +1,4 @@
 addSbtPlugin("com.typesafe.sbt" % "sbt-scalariform" % "1.2.0")
+
+addSbtPlugin("com.typesafe.sbt" % "sbt-osgi" % "0.7.0")
+
