+++ spray-websocket/src/main/scala/spray/can/websocket/server/UpgradeSupport.scala
